pyproject.toml
[project]
name = "yournextrepresentative"
version = "0.0.0"
description = "A website for crowd-sourcing structured election candidate data"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "amazon-textract-textractor[pdf]==1.9.2",
    "beautifulsoup4==4.12.0",
    "blessed==1.20.0", # needed by django-q2
    "croniter==6.0.0", # needed by django-q2
    "nh3==0.2.21",
    "boto3==1.34.105",
    "django-cors-headers==4.9.0",
    "django-date-extensions==3.1.2",
    "python-dotenv==1.0.1",
    "django-extensions==4.1",
    "django-filter==25.1",
    "django-hCaptcha==0.2.0",
    "django-model-utils==5.0.0",
    "django-pipeline==4.1.0", # TODO: remove when we migrate to dc_django_utils
    "django-q2==1.8.0",
    "django-q-registry==0.5.0",
    "django-sesame==3.2.3",
    "django-storages==1.14.6",
    "django-svelte==0.2.1",
    "django==5.2.9",
    "djangorestframework-jsonp==1.0.2",
    "djangorestframework-csv==3.0.2",
    "djangorestframework==3.16.1",
    "drf-yasg==1.21.11",
    "uk-election-timetables==4.2.3",
    "uk-election-ids==0.9.1",
    "gender-detector==0.1.0",
    "gunicorn==23.0.0",
    "jsonpatch==1.32",
    "jsonpointer==2.3",
    "libsass==0.22.0", # TODO: remove when we migrate to dc_django_utils
    "lxml==6.0.2",
    "Markdown==3.4.3",
    "markdown-it-py==4.0.0",
    "nameparser==1.1.2",
    "ndg-httpsclient==0.5.1",
    "openai==1.30.3",
    # for compatibility with openai==1.30.3
    # TODO: review/remove when we upgrade openai package
    "httpx==0.27.0",
    "Pillow==10.3.0",
    "psycopg==3.1.12",
    "python-dateutil==2.8.2",
    "python-magic==0.4.27",
    "python-slugify==8.0.1",
    "requests==2.32.4",
    "sentry-sdk[django]==2.18.0",
    "sorl-thumbnail==12.11.0",
    "whitenoise==6.5.0",
    "sorl-thumbnail-serializer-field",
    "slacker2",
    # SOPN parsing
    "pdfminer.six==20201018",
    "pypandoc_binary==1.14",
    "PyPDF2==2.12.1",
    "amazon-textract-response-parser==1.0.3",
    "amazon-textract-helper==0.0.35",
    "pandas>=3.0.0",
]

[dependency-groups]
cdk = [
    "aws-cdk-lib==2.178.0",
    "constructs==10.4.2",
    "pytest==7.2.2",
]
production = [
    "aws-xray-sdk==2.14.0",
]
dev = [
    "django-debug-toolbar-template-timings==0.9",
    "django-debug-toolbar==4.2.0",
    "django-webtest==1.9.10",
    "factory-boy==3.2.1",
    "Faker==19.2.0",
    "freezegun==1.2.2",
    "ipdb==0.13.13",
    "ipython==8.12.3",
    "mock==5.0.1", # TODO: bin and replace with unittest.mock
    "moto[s3]==4.2.11",
    "pytest==7.2.2",
    "pytest-dotenv==0.5.2",
    "pytest-cov==4.0.0",
    "pytest-django==4.5.2",
    "pytest-freezegun==0.4.2",
    "pytest-subtests==0.13.1",
    "pytest-ruff==0.2.1",
    "responses==0.25.0",
    "ruff==0.2.2",
    "watchdog==4.0.2",
    "WebTest==3.0.0",
]

[tool.uv]
package = false
required-version = "==0.9.*"

[tool.uv.sources]
sorl-thumbnail-serializer-field = { git = "https://github.com/dessibelle/sorl-thumbnail-serializer-field.git", tag = "0.2" }
slacker2 = { git = "https://github.com/symroe/slacker.git", rev = "447cc8ca" }

[tool.ruff]
line-length = 80
lint.ignore = ["E501"]
lint.extend-select = [
    "I",
    "C4",
    "SIM",
    "Q003",
    "RET",
]
extend-exclude = ["ynr/settings/local.py"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "@abstractmethod"
]

[tool.pytest.ini_options]
norecursedirs =[
    "ynr/static",
    "ynr/media",
    "deploy",
    ".aws-sam",
    ".ruff-cache",
    "locale",
    ".git",
    "data",
    "docs",
    "templates",
    ".download-cache",
    "__pycache__",

]
addopts =[
    "--ignore=*/__pycache__",
    "--ignore=.*",
    "--ignore=.download-cache",
    "--ignore=democracy_club/settings/local.py",
    "--ignore=lib",
    "--ignore=cdk",
    "--ignore=node_modules",
    "--ruff",
    "--ds=ynr.settings.testing"
]

FAIL_INVALID_TEMPLATE_VARS = 1

env_files = [".env"]
